+++ player.py
@@ -1,0 +1,187 @@
+import pygame
+from elements import SpriteSheet
+from settings import SCREEN_HEIGHT
+
+
+
+# ==========================================================================
+#                     PLAYER OBJECT
+# ==========================================================================
+class Player(pygame.sprite.Sprite):
+    """
+    This class represents the bar at the bottom that the player controls.
+    """
+ 
+    # -- Methods
+    def __init__(self, height=40, width=60, color=(255, 0, 0)):
+        """ Constructor function """
+
+        # Call the parent's constructor
+        super().__init__()
+ 
+        # Create an image of the block, and fill it with a color.
+        # This could also be an image loaded from the disk.
+        self.width = width
+        self.height = height
+
+        self.walking_frames_l = []
+        self.walking_frames_r = []
+
+
+        # self.image = pygame.Surface([self.width, self.height])
+        # self.image.fill(color)
+        self.direction = "R"
+
+
+        sprite_sheet = SpriteSheet("images/p1_walk.png")
+
+
+        image = sprite_sheet.get_image(0, 0, 66, 90)
+        self.walking_frames_r.append(image)
+        image = sprite_sheet.get_image(66, 0, 66, 90)
+        self.walking_frames_r.append(image)
+        image = sprite_sheet.get_image(132, 0, 67, 90)
+        self.walking_frames_r.append(image)
+        image = sprite_sheet.get_image(0, 93, 66, 90)
+        self.walking_frames_r.append(image)
+        image = sprite_sheet.get_image(66, 93, 66, 90)
+        self.walking_frames_r.append(image)
+        image = sprite_sheet.get_image(132, 93, 72, 90)
+        self.walking_frames_r.append(image)
+        image = sprite_sheet.get_image(0, 186, 70, 90)
+        self.walking_frames_r.append(image)
+ 
+        # Load all the right facing images, then flip them
+        # to face left.
+        image = sprite_sheet.get_image(0, 0, 66, 90)
+        image = pygame.transform.flip(image, True, False)
+        self.walking_frames_l.append(image)
+        image = sprite_sheet.get_image(66, 0, 66, 90)
+        image = pygame.transform.flip(image, True, False)
+        self.walking_frames_l.append(image)
+        image = sprite_sheet.get_image(132, 0, 67, 90)
+        image = pygame.transform.flip(image, True, False)
+        self.walking_frames_l.append(image)
+        image = sprite_sheet.get_image(0, 93, 66, 90)
+        image = pygame.transform.flip(image, True, False)
+        self.walking_frames_l.append(image)
+        image = sprite_sheet.get_image(66, 93, 66, 90)
+        image = pygame.transform.flip(image, True, False)
+        self.walking_frames_l.append(image)
+        image = sprite_sheet.get_image(132, 93, 72, 90)
+        image = pygame.transform.flip(image, True, False)
+        self.walking_frames_l.append(image)
+        image = sprite_sheet.get_image(0, 186, 70, 90)
+        image = pygame.transform.flip(image, True, False)
+        self.walking_frames_l.append(image)
+
+
+
+        self.image = self.walking_frames_r[0]
+        # Set a referance to the image rect.
+        self.rect = self.image.get_rect()
+ 
+        # Set speed vector of player
+        self.change_x = 0
+        self.change_y = 0
+
+            
+        # List of sprites we can bump against
+        self.level = None
+        self.player_over_threshold = False
+        self.died = False
+
+
+
+    def __str__(self):
+        message = "Player: \n"
+        message += "\theight: {}\n".format(self.height)
+        message += "\twidth: {}\n".format(self.width)
+        message += "\tpos_x: {}\n".format(self.rect.x)
+        message += "\tpos_y: {}\n".format(self.rect.y)
+        return message
+    
+    def update(self):
+        """ Move the player. """
+        # Gravity
+        self.calc_grav()
+ 
+        # Move left/right
+        self.rect.x += self.change_x
+
+        pos = self.rect.x + self.level.world_shift
+        if self.direction == "R":
+            frame = (pos // 30) % len(self.walking_frames_r)
+            self.image = self.walking_frames_r[frame]
+        else:
+            frame = (pos // 30) % len(self.walking_frames_l)
+            self.image = self.walking_frames_l[frame]
+
+        
+        # See if we hit anything
+        block_hit_list = pygame.sprite.spritecollide(self, self.level.platform_list, False)
+        for block in block_hit_list:
+            # If we are moving right,
+            # set our right side to the left side of the item we hit
+            if self.change_x > 0:
+                self.rect.right = block.rect.left
+            elif self.change_x < 0:
+                # Otherwise if we are moving left, do the opposite.
+                self.rect.left = block.rect.right
+ 
+        # Move up/down
+        self.rect.y += self.change_y
+ 
+        # Check and see if we hit anything
+        block_hit_list = pygame.sprite.spritecollide(self, self.level.platform_list, False)
+        for block in block_hit_list:
+ 
+            # Reset our position based on the top/bottom of the object.
+            if self.change_y > 0:
+                self.rect.bottom = block.rect.top
+            elif self.change_y < 0:
+                self.rect.top = block.rect.bottom
+ 
+            # Stop our vertical movement
+            self.change_y = 0
+ 
+    def calc_grav(self):
+        """ Calculate effect of gravity. """
+        if self.change_y == 0:
+            self.change_y = 1
+        else:
+            self.change_y += .35
+ 
+        # See if we are on the ground.
+        if self.player_over_threshold == False:
+            if self.rect.y >= SCREEN_HEIGHT - self.rect.height and self.change_y >= 0:
+                self.change_y = 0
+                self.rect.y = SCREEN_HEIGHT - self.rect.height
+
+    def jump(self):
+        """ Called when user hits 'jump' button. """
+ 
+        # move down a bit and see if there is a platform below us.
+        # Move down 2 pixels because it doesn't work well if we only move down 1
+        # when working with a platform moving down.
+        self.rect.y += 2
+        platform_hit_list = pygame.sprite.spritecollide(self, self.level.platform_list, False)
+        self.rect.y -= 2
+ 
+        # If it is ok to jump, set our speed upwards
+        if len(platform_hit_list) > 0 or self.rect.bottom >= SCREEN_HEIGHT:
+            self.change_y = -10
+ 
+    # Player-controlled movement:
+    def go_left(self):
+        """ Called when the user hits the left arrow. """
+        self.change_x = -6
+        self.direction = "L"
+    def go_right(self):
+        """ Called when the user hits the right arrow. """
+        self.change_x = 6
+        self.direction = "R"
+    def stop(self):
+        """ Called when the user lets off the keyboard. """
+        self.change_x = 0
+
