+++ settings.py
@@ -16,11 +16,10 @@ SCREEN_WIDTH = 800
 SCREEN_HEIGHT = 600
 
 #Frame rate
-FPS = 60
+FPS = 100
 
 # PLAYER COLOR
 PLAYER_COLOR = ORANGE
-
 
 # LEVEL COLOR
 LEVEL_COLOR = LIGHT_CYAN
@@ -29,21 +28,16 @@ LEVEL_COLOR = LIGHT_CYAN
 PLAYER_HEIGHT = 80
 PLAYER_WIDTH = 60
 
-
 # PLATFORM COLOR 
 PLATFORM_COLOR = DARK_GREEN
 PLATFORM_HEIGHT = PLAYER_HEIGHT - 10
 
-
-
-
 # WINDOW BAR TITLE
 WINDOW_TITLE = "New little game"
 
-
+# GAME LOOP SETTINGS
 HIGHEST_JUMP_RATE = 10
 HIGHEST_JUMP_THRESHOLD = 300
 
-
-
+# LEVEL SPRITE SHEET
 SPRITE_SHEET_PATH = "images/tiles_spritesheet.png"
